edit solution for 2-

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -9,7 +9,7 @@
 
 void *_calloc(unsigned int nmemb, unsigned int size)
 {
-	int *p;
+	double *p;
 	unsigned int i;
 
 	if (nmemb <= 0 || size <= 0)
